improve coverage xtask
