+++ xtask/Cargo.toml
@@ -7,5 +7,7 @@ edition = "2018"
 [dependencies]
 anyhow = "1"
 json = "0.12"
+once_cell = "1"
 pico-args = "0.3"
+regex = "1.4"
 wait-timeout = "0.2"
